Moving old Maya scripts to old folder; shelf is back to its original state
