+++ maya-tools/shelf/scripts/maya_save_version.py
@@ -1,19 +1,24 @@
 import maya.cmds as cmds
 import os
 import utilities as amu
-import maya_checkout
+import maya_checkout # This isn't used... probably because the checkout module needs context, and this doesn't have it.
 import maya_checkin
+import Facade.facade as facade
 
 def go():
     filePath = cmds.file(q=True, sceneName=True)
-    toCheckin = os.path.join(amu.getUserCheckoutDir(), os.path.basename(os.path.dirname(filePath)))
-    toCheckout = amu.getCheckinDest(toCheckin) 
+    # Need to grab the asset name and asset type so we can check it out later on.
+    assetType = facade.getAssetType(filePath)
+    assetName = facade.getAssetName(filePath)
+    # print "assetType: ", assetType
+    # print "assetName: ", assetName
     maya_checkin.checkin()
+
     try:
-        destpath = amu.checkout(toCheckout, True)
+        destpath = facade.checkout(assetName, True, assetType)
     except Exception as e:
         print str(e)
-        if not amu.checkedOutByMe(toCheckout):
+        if not facade.checkedOutByMe(assetName, assetType):
             cmd.confirmDialog(  title          = 'Can Not Checkout'
                                , message       = str(e)
                                , button        = ['Ok']
@@ -22,8 +27,10 @@ def go():
                                , dismissString = 'Ok')
             return
         else:
-            destpath = amu.getCheckoutDest(toCheckout)
-    filename = os.path.basename(os.path.dirname(toCheckout))+'_'+os.path.basename(toCheckout)+'.mb'
+            destpath = facade.getCheckoutDest(assetName, assetType)
+
+    filename = facade.getFilepath(destpath, assetName, assetType)
+    filename = filename + ".mb"
     toOpen = os.path.join(destpath, filename)
     # open the file
     if os.path.exists(toOpen):
